+++ src/contraction/src/contract_function.cpp
@@ -7,13 +7,13 @@ using namespace std;
 using namespace boost;
 typedef adjacency_list<vecS, vecS, undirectedS, Vertex,Edge> G;
 //calculates the shortest path,returns the pathsize and also the path as an array of edges
-int fetch_contracted_graph(Edge *edges,int *initial_num_vertices,
-							int *final_num_vertices,int initial_num_edges,
-							int *final_num_edges,int level,char **graphName,
+int64_t fetch_contracted_graph(Edge *edges,int64_t *initial_num_vertices,
+							int64_t *final_num_vertices,int64_t initial_num_edges,
+							int64_t *final_num_edges,int64_t level,char **graphName,
 							char **edgeString,char **psuedoEString,
 							char **removedVString,char **removedEString)
 {
-	int final_edge_count=0;
+	int64_t final_edge_count=0;
 	typedef Graph_Minimizer<G> Graph;
 	Graph g(UNDIRECTED,*initial_num_vertices);
 	g.initialize_graph_minimizer(edges,initial_num_edges);
