+++ src/contraction/src/dijkstra_function.h
@@ -7,9 +7,9 @@
 extern "C" {
 #endif
 
-int compute_dijkstra(int64_t source,int64_t target,Edge *edges,int64_t num_vertices,int64_t num_edges,Edge **path);
+int64_t compute_dijkstra(int64_t source,int64_t target,Edge *edges,int64_t num_vertices,int64_t num_edges,Edge **path);
 
-int dijkstra_on_contracted(int64_t source,int64_t target,
+int64_t dijkstra_on_contracted(int64_t source,int64_t target,
 							pgr_contracted_blob **graphInfo,bool has_rcost);
 #ifdef __cplusplus
 }
